Move FAQ commands to a group
